Get runc to build clean on Solaris
